move a lot of things to Sound.Tidal.UI, port over some more functions from master, rejig `_chop`
